Handle joins in arithmetic expressions (#1765)
* Move query reuse tests to test_query_reuse.py

* Add test exposing issue

* Return joins from CombinedExpression.resolve
